requirements.txt
# Core Dependencies
flask==3.0.0
flask-cors==4.0.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
python-dotenv==1.0.0
psutil==5.9.5
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
colorama==0.4.6
